pyproject.toml
[tool.poetry]
name = "ld4lg"
version = "1.0.0"
description = "ld4lg"
authors = ["Koki Hokao"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
accelerate = "*"
datasets = "*"
einops = "*"
ema_pytorch = "*"
evaluate = "*"
hydra-core = "*"
mauve-text = "*"
nltk = "*"
numpy = "*"
spacy = "*"
sentence_transformers = "*"
torch = "==1.12.1"
tqdm = "*"
transformers = "*"
wandb = "*"

[tool.poetry.dev-dependencies]
autopep8 = "*"
flake8 = "*"
flake8-isort = "*"
flake8-quotes = "*"
jupyter = "*"
pep8-naming = "*"
pyproject-flake8 = "*"
seaborn = "*"

[tool.isort]
line_length = 120
multi_line_output = 3

[tool.flake8]
max-line-length = 120
exclude = "*/__init__.py"
ignore = "W503,N803"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
